% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gifrecord.R
\name{gifrecord-shiny}
\alias{gifrecord-shiny}
\alias{gifrecordOutput}
\alias{renderGifrecord}
\title{Shiny bindings for gifrecord}
\usage{
gifrecordOutput(outputId, width = "260px", height = "2.5em")

renderGifrecord(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width,height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a gifrecord}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
  is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using gifrecord within Shiny
applications and interactive Rmd documents.
}

