% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ gifrecordeR }
\alias{ gifrecordeR }
\alias{ gifrecordeR }
\title{ gifrecordeR }
\description{
Easily record animated GIFs of your web session straight from R with this htmlwidget wrapper for \href{https://github.com/yaronn/gifw00t/}{gifw00t}.
}
