% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gifrecord.R
\name{gifrecord}
\alias{gifrecord}
\title{Add a GIF Recorder}
\usage{
gifrecord(message)
}
\description{
Record animated GIFs of your browser, RStudio viewer, or
 Shiny app with the convenient infrastructure of an
 htmlwidget thanks to \href{https://github.com/yaronn/gifw00t/}{gifw00t}.
}

